Update gems and tests to work with Yolk
This updates several of the test cases (and corresponding VCR cassettes)
to work with the latest Yolk.  The test suite modifies Yolk and cannot
be run back to back without resetting the Yolk database.

The db/dump directory contains a copy of the Yolk database needed to run
this test suite.

Some tests still fail occassionally when run with VCR cassettes.